Moved UI classes to their own package
